+++ src/Other/Strategies/Continuous.php
@@ -4,5 +4,17 @@ namespace Rubix\ML\Other\Strategies;
 
 interface Continuous extends Strategy
 {
-    //
+    /**
+     * Fit the guessing strategy to a set of values.
+     *
+     * @param (int|float)[] $values
+     */
+    public function fit(array $values) : void;
+
+    /**
+     * Make a guess.
+     *
+     * @return int|float
+     */
+    public function guess();
 }