+++ src/Other/Strategies/WildGuess.php
@@ -10,8 +10,8 @@ use const Rubix\ML\PHI;
 /**
  * Wild Guess
  *
- * Guess a random number somewhere between an upper and lower bound given by
- * the data and a user-defined *shrinkage* parameter.
+ * Guess a random number somewhere between the minimum and maximum computed by fitting a
+ * collection of values.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -20,46 +20,23 @@ use const Rubix\ML\PHI;
 class WildGuess implements Continuous
 {
     /**
-     * The range between the upper and lower bounds of the guess. A value of 1.0
-     * indicates the full range of fitted values, whereas the range becomes narrower
-     * as the parameter goes to 0.
-     *
-     * @var float
-     */
-    protected $alpha;
-
-    /**
      * The minimum value of the fitted data.
      *
-     * @var float|null
+     * @var int|null
      */
     protected $min;
 
     /**
      * The maximum value of the fitted data.
      *
-     * @var float|null
+     * @var int|null
      */
     protected $max;
 
     /**
-     * @param float $alpha
-     * @throws \InvalidArgumentException
-     */
-    public function __construct(float $alpha = 0.5)
-    {
-        if ($alpha <= 0. or $alpha > 1.) {
-            throw new InvalidArgumentException('Alpha must be between'
-                . " 0 and 1, $alpha given.");
-        }
-
-        $this->alpha = $alpha;
-    }
-
-    /**
      * Fit the guessing strategy to a set of values.
      *
-     * @param (string|int|float)[] $values
+     * @param (int|float)[] $values
      * @throws \InvalidArgumentException
      */
     public function fit(array $values) : void
@@ -69,18 +46,8 @@ class WildGuess implements Continuous
                 . ' to at least 1 value.');
         }
 
-        $this->min = $this->alpha * min($values);
-        $this->max = $this->alpha * max($values);
-    }
-
-    /**
-     * Return the lower and upper bounds in a 2-tuple.
-     *
-     * @return (int|float|null)[]
-     */
-    public function range() : array
-    {
-        return [$this->min, $this->max];
+        $this->min = (int) round(min($values) * PHI);
+        $this->max = (int) round(max($values) * PHI);
     }
 
     /**
@@ -95,9 +62,6 @@ class WildGuess implements Continuous
             throw new RuntimeException('Strategy has not been fitted.');
         }
 
-        $min = (int) round($this->min * PHI);
-        $max = (int) round($this->max * PHI);
-
-        return rand($min, $max) / PHI;
+        return rand($this->min, $this->max) / PHI;
     }
 }