+++ tests/NeuralNet/Layers/DropoutTest.php
@@ -82,9 +82,9 @@ class DropoutTest extends TestCase
         $this->layer->initialize($this->fanIn);
 
         $expected = [
-            [0., 0., -0.],
-            [0., 0., 0.],
-            [0.002, -6.0, -0.5],
+            [1.0, 2.5, -0.1],
+            [0.1, 0., 3.],
+            [0.002, -6.0, 0.],
         ];
 
         $forward = $this->layer->forward($this->input);
@@ -95,9 +95,9 @@ class DropoutTest extends TestCase
         $gradient = $this->layer->back($this->prevGrad, $this->optimizer)->compute();
 
         $expected = [
-            [0., 0., 0.],
-            [0., 0.2, 0.],
-            [0.25, 0.1, 0.89],
+            [0.25, 0.7, 0.1],
+            [0.5, 0.2, 0.01],
+            [0.25, 0.1, 0.0],
         ];
 
         $this->assertInstanceOf(Matrix::class, $gradient);